workers/timescaledb-event-etl/requirements.txt
aiokafka==0.11.0
orjson==3.10.7
psycopg2-binary==2.9.9
prometheus-client==0.20.0
